convert snippets
