+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/DeserializeFromFile.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.IO;
+using System.Text.Json;
+
+namespace DeserializeFromFile
+{
+    public class Program
+    {
+        public class WeatherForecast
+        {
+            public DateTimeOffset Date { get; set; }
+            public int TemperatureCelsius { get; set; }
+            public string Summary { get; set; }
+        }
+
+        public static void Main()
+        {
+            string fileName = "WeatherForecast.json";
+            string jsonString = File.ReadAllText(fileName);
+            WeatherForecast weatherForecast = JsonSerializer.Deserialize<WeatherForecast>(jsonString);
+
+            Console.WriteLine($"Date: {weatherForecast.Date}");
+            Console.WriteLine($"TemperatureCelsius: {weatherForecast.TemperatureCelsius}");
+            Console.WriteLine($"Summary: {weatherForecast.Summary}");
+        }
+    }
+}
+// output:
+//Date: 8/1/2019 12:00:00 AM -07:00
+//TemperatureCelsius: 25
+//Summary: Hot
