+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/DeserializeFromFileAsync.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.IO;
+using System.Text.Json;
+using System.Threading.Tasks;
+
+namespace DeserializeFromFileAsync
+{
+    public class Program
+    {
+        public class WeatherForecast
+        {
+            public DateTimeOffset Date { get; set; }
+            public int TemperatureCelsius { get; set; }
+            public string Summary { get; set; }
+        }
+        public static async Task Main()
+        {
+            string fileName = "WeatherForecast.json";
+            using FileStream openStream = File.OpenRead(fileName);
+            WeatherForecast weatherForecast = 
+                await JsonSerializer.DeserializeAsync<WeatherForecast>(openStream);
+
+            Console.WriteLine($"Date: {weatherForecast.Date}");
+            Console.WriteLine($"TemperatureCelsius: {weatherForecast.TemperatureCelsius}");
+            Console.WriteLine($"Summary: {weatherForecast.Summary}");
+        }
+    }
+}
+// output:
+//Date: 8/1/2019 12:00:00 AM -07:00
+//TemperatureCelsius: 25
+//Summary: Hot
