--- docs/standard/serialization/snippets/system-text-json-how-to/csharp/RoundtripToFile.cs
@@ -1,68 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Text.Json;
-
-namespace RoundtripToFile1
-{
-    // <Serialize>
-    public class Program
-    {
-        public class WeatherForecast
-        {
-            public DateTimeOffset Date { get; set; }
-            public int TemperatureCelsius { get; set; }
-            public string Summary { get; set; }
-        }
-
-        public static void Main()
-        {
-            var weatherForecast = new WeatherForecast
-            {
-                Date = DateTime.Parse("2019-08-01"),
-                TemperatureCelsius = 25,
-                Summary = "Hot"
-            };
-
-            string fileName = "WeatherForecast.json"; 
-            string jsonString = JsonSerializer.Serialize(weatherForecast);
-            File.WriteAllText(fileName, jsonString);
-
-            Console.WriteLine(File.ReadAllText(fileName));
-        }
-    }
-    // output:
-    //{"Date":"2019-08-01T00:00:00-07:00","TemperatureCelsius":25,"Summary":"Hot"}
-    // </Serialize>
-}
-
-
-
-namespace RoundtripToFile2
-{
-    // <Deserialize>
-    public class Program
-    {
-        public class WeatherForecast
-        {
-            public DateTimeOffset Date { get; set; }
-            public int TemperatureCelsius { get; set; }
-            public string Summary { get; set; }
-        }
-
-        public static void Main()
-        {
-            string fileName = "WeatherForecast.json";
-            string jsonString = File.ReadAllText(fileName);
-            WeatherForecast weatherForecast = JsonSerializer.Deserialize<WeatherForecast>(jsonString);
-
-            Console.WriteLine($"Date: {weatherForecast.Date}");
-            Console.WriteLine($"TemperatureCelsius: {weatherForecast.TemperatureCelsius}");
-            Console.WriteLine($"Summary: {weatherForecast.Summary}");
-        }
-    }
-    // output:
-    //Date: 8/1/2019 12:00:00 AM -07:00
-    //TemperatureCelsius: 25
-    //Summary: Hot
-    // </Deserialize>
-}
