--- docs/standard/serialization/snippets/system-text-json-how-to/csharp/RoundtripToFileAsync.cs
@@ -1,70 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Text.Json;
-using System.Threading.Tasks;
-
-namespace RoundtripToFileAsync1
-{
-    // <Serialize>
-    public class Program
-    {
-        public class WeatherForecast
-        {
-            public DateTimeOffset Date { get; set; }
-            public int TemperatureCelsius { get; set; }
-            public string Summary { get; set; }
-        }
-
-        public static async Task Main()
-        {
-            var weatherForecast = new WeatherForecast
-            {
-                Date = DateTime.Parse("2019-08-01"),
-                TemperatureCelsius = 25,
-                Summary = "Hot"
-            };
-
-            string fileName = "WeatherForecast.json";
-            using FileStream createStream = File.Create(fileName);
-            await JsonSerializer.SerializeAsync(createStream, weatherForecast);
-            await createStream.DisposeAsync();
-
-            Console.WriteLine(File.ReadAllText(fileName));
-        }
-    }
-    // output:
-    //{"Date":"2019-08-01T00:00:00-07:00","TemperatureCelsius":25,"Summary":"Hot"}
-    // </Serialize>
-}
-
-
-
-namespace RoundtripToFileAsync2
-{
-    // <Deserialize>
-    public class Program
-    {
-        public class WeatherForecast
-        {
-            public DateTimeOffset Date { get; set; }
-            public int TemperatureCelsius { get; set; }
-            public string Summary { get; set; }
-        }
-        public static async Task Main()
-        {
-            string fileName = "WeatherForecast.json";
-            using FileStream openStream = File.OpenRead(fileName);
-            WeatherForecast weatherForecast = 
-                await JsonSerializer.DeserializeAsync<WeatherForecast>(openStream);
-
-            Console.WriteLine($"Date: {weatherForecast.Date}");
-            Console.WriteLine($"TemperatureCelsius: {weatherForecast.TemperatureCelsius}");
-            Console.WriteLine($"Summary: {weatherForecast.Summary}");
-        }
-    }
-    // output:
-    //Date: 8/1/2019 12:00:00 AM -07:00
-    //TemperatureCelsius: 25
-    //Summary: Hot
-    // </Deserialize>
-}
