+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/RoundtripToUtf8.cs
@@ -23,8 +23,8 @@ namespace RoundtripToUtf8Bytes1
                 Summary = "Hot"
             };
 
-            byte[] jsonUtf8Bytes;
-            jsonUtf8Bytes = JsonSerializer.SerializeToUtf8Bytes(weatherForecast);
+            byte[] jsonUtf8Bytes =
+                JsonSerializer.SerializeToUtf8Bytes(weatherForecast);
 
             Console.WriteLine(Encoding.UTF8.GetString(jsonUtf8Bytes));
         }
@@ -36,7 +36,6 @@ namespace RoundtripToUtf8Bytes1
 
 namespace RoundtripToUtf8Bytes2
 {
-    // <Deserialize1>
     public class Program
     {
         public record WeatherForecast
@@ -55,11 +54,14 @@ namespace RoundtripToUtf8Bytes2
                 Summary = "Hot"
             };
 
-            // Get JSON in a ReadOnlySpan<byte> for deserialization example.
-            var readOnlySpan = new ReadOnlySpan<byte>(JsonSerializer.SerializeToUtf8Bytes(weatherForecast));
+            byte[] jsonUtf8Bytes =
+                JsonSerializer.SerializeToUtf8Bytes(weatherForecast);
 
+            // <Deserialize1>
+            var readOnlySpan = new ReadOnlySpan<byte>(jsonUtf8Bytes);
             WeatherForecast deserializedWeatherForecast = 
                 JsonSerializer.Deserialize<WeatherForecast>(readOnlySpan);
+            // </Deserialize1>
 
             Console.WriteLine($"Date: {deserializedWeatherForecast.Date}");
             Console.WriteLine($"TemperatureCelsius: {deserializedWeatherForecast.TemperatureCelsius}");
@@ -70,12 +72,10 @@ namespace RoundtripToUtf8Bytes2
     //Date: 8/1/2019 12:00:00 AM -07:00
     //TemperatureCelsius: 25
     //Summary: Hot
-    // </Deserialize1>
 }
 
 namespace RoundtripToUtf8Bytes3
 {
-    // <Deserialize2>
     public class Program
     {
         public record WeatherForecast
@@ -94,10 +94,13 @@ namespace RoundtripToUtf8Bytes3
                 Summary = "Hot"
             };
 
-            // Get JSON in a Utf8JsonReader for deserialization example.
-            var utf8Reader = new Utf8JsonReader(JsonSerializer.SerializeToUtf8Bytes(weatherForecast));
+            byte[] jsonUtf8Bytes = JsonSerializer.SerializeToUtf8Bytes(weatherForecast);
 
-            WeatherForecast deserializedWeatherForecast = JsonSerializer.Deserialize<WeatherForecast>(ref utf8Reader);
+            // <Deserialize2>
+            var utf8Reader = new Utf8JsonReader(jsonUtf8Bytes);
+            WeatherForecast deserializedWeatherForecast = 
+                JsonSerializer.Deserialize<WeatherForecast>(ref utf8Reader);
+            // </Deserialize2>
 
             Console.WriteLine($"Date: {deserializedWeatherForecast.Date}");
             Console.WriteLine($"TemperatureCelsius: {deserializedWeatherForecast.TemperatureCelsius}");
@@ -108,5 +111,4 @@ namespace RoundtripToUtf8Bytes3
     //Date: 8/1/2019 12:00:00 AM -07:00
     //TemperatureCelsius: 25
     //Summary: Hot
-    // </Deserialize2>
 }
