+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/SerializeBasic.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Text.Json;
+
+namespace SerializeBasic
+{
+    public class Program
+    {
+        public class WeatherForecast
+        {
+            public DateTimeOffset Date { get; set; }
+            public int TemperatureCelsius { get; set; }
+            public string Summary { get; set; }
+        }
+
+        public static void Main()
+        {
+            var weatherForecast = new WeatherForecast
+            {
+                Date = DateTime.Parse("2019-08-01"),
+                TemperatureCelsius = 25,
+                Summary = "Hot"
+            };
+
+            string jsonString = JsonSerializer.Serialize(weatherForecast);
+
+            Console.WriteLine(jsonString);
+        }
+    }
+}
+// output:
+//{"Date":"2019-08-01T00:00:00-07:00","TemperatureCelsius":25,"Summary":"Hot"}
