+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/SerializeToFile.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.IO;
+using System.Text.Json;
+
+namespace SerializeToFile
+{
+    public class Program
+    {
+        public class WeatherForecast
+        {
+            public DateTimeOffset Date { get; set; }
+            public int TemperatureCelsius { get; set; }
+            public string Summary { get; set; }
+        }
+
+        public static void Main()
+        {
+            var weatherForecast = new WeatherForecast
+            {
+                Date = DateTime.Parse("2019-08-01"),
+                TemperatureCelsius = 25,
+                Summary = "Hot"
+            };
+
+            string fileName = "WeatherForecast.json"; 
+            string jsonString = JsonSerializer.Serialize(weatherForecast);
+            File.WriteAllText(fileName, jsonString);
+
+            Console.WriteLine(File.ReadAllText(fileName));
+        }
+    }
+}
+// output:
+//{"Date":"2019-08-01T00:00:00-07:00","TemperatureCelsius":25,"Summary":"Hot"}
