+++ docs/standard/serialization/snippets/system-text-json-how-to/csharp/SerializeWriteIndented.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Text.Json;
+
+namespace SerializeWriteIndented
+{
+    public class Program
+    {
+        public class WeatherForecast
+        {
+            public DateTimeOffset Date { get; set; }
+            public int TemperatureCelsius { get; set; }
+            public string Summary { get; set; }
+        }
+
+        public static void Main()
+        {
+            var weatherForecast = new WeatherForecast
+            {
+                Date = DateTime.Parse("2019-08-01"),
+                TemperatureCelsius = 25,
+                Summary = "Hot"
+            };
+
+            var options = new JsonSerializerOptions { WriteIndented = true };
+            string jsonString = JsonSerializer.Serialize(weatherForecast, options);
+            Console.WriteLine(jsonString);
+        }
+    }
+}
+// output:
+//{
+//  "Date": "2019-08-01T00:00:00-07:00",
+//  "TemperatureCelsius": 25,
+//  "Summary": "Hot"
+//}
